API_PyTest/requirements.txt
mysql-connector-python==8.1.0
pytest==7.4.2
pytest-steps==1.8.0
Random-Word==1.0.11
RandomWords==0.4.0
requests==2.31.0
